+++ examples/sandbox/sandbox.py
@@ -9,14 +9,16 @@ netParams.sizeZ = 100 # z-dimension (horizontal length) size in um
 netParams.propVelocity = 100.0 # propagation velocity (um/ms)
 netParams.probLengthConst = 150.0 # length constant for conn probability (um)
 
-
+import json
+file = 'ssc-3_spikes.json'
+with open(file, 'r') as f: spks = json.load(f)
+print spks
 ## Population parameters
-netParams.popParams['E2'] = {'cellType': 'E', 'numCells': 3, 'yRange': [100,300], 'cellModel': 'HH'}
-netParams.popParams['I2'] = {'cellType': 'I', 'numCells': 3, 'yRange': [100,300], 'cellModel': 'HH'}
-netParams.popParams['E4'] = {'cellType': 'E', 'numCells': 3, 'yRange': [300,600], 'cellModel': 'HH'}
-netParams.popParams['I4'] = {'cellType': 'I', 'numCells': 3, 'yRange': [300,600], 'cellModel': 'HH'}
-netParams.popParams['E5'] = {'cellType': 'E', 'numCells': 3, 'ynormRange': [0.6,1.0], 'cellModel': 'HH'}
-netParams.popParams['I5'] = {'cellType': 'I', 'numCells': 3, 'ynormRange': [0.6,1.0], 'cellModel': 'HH'}
+#netParams.popParams['E2'] = {'cellType': 'E', 'numCells': 3, 'yRange': [100,300], 'cellModel': 'HH'}
+netParams.popParams['S2'] = {'cellModel': 'VecStim', 'numCells': 1000, 'spkTimes': spks} #[[50, 100, 200, 300], [150, 240, 412, 320],[55, 105, 210, 330]] }
+#netParams.popParams['S3'] = {'cellModel': 'VecStim', 'numCells': 3, 'cellsList': [{'spkTimes': [50, 100, 200, 300]},
+																				# {'spkTimes': [150, 240, 412, 320]},
+																				# {'spkTimes': [55, 105, 210, 330]}]}
 
 
 ## Cell property rules
@@ -24,57 +26,45 @@ cellRule = {'conds': {'cellType': 'E'},  'secs': {}}  # cell rule dict
 cellRule['secs']['soma'] = {'geom': {}, 'mechs': {}}                              # soma params dict
 cellRule['secs']['soma']['geom'] = {'diam': 15, 'L': 14, 'Ra': 120.0}                   # soma geometry
 cellRule['secs']['soma']['mechs']['hh'] = {'gnabar': 0.13, 'gkbar': 0.036, 'gl': 0.003, 'el': -70}      # soma hh mechanism
-netParams.cellParams['Erule'] = cellRule                          # add dict to list of cell params
 
-cellRule = {'conds': {'cellType': 'I'},  'secs': {}}  # cell rule dict
-cellRule['secs']['soma'] = {'geom': {}, 'mechs': {}}                              # soma params dict
-cellRule['secs']['soma']['geom'] = {'diam': 10.0, 'L': 9.0, 'Ra': 110.0}                  # soma geometry
-cellRule['secs']['soma']['mechs']['hh'] = {'gnabar': 0.11, 'gkbar': 0.036, 'gl': 0.003, 'el': -70}      # soma hh mechanism
-netParams.cellParams['Irule'] = cellRule                          # add dict to list of cell params
+netParams.cellParams['Erule'] = cellRule                          # add dict to list of cell params
 
 
 ## Synaptic mechanism parameters
 netParams.synMechParams['exc'] = {'mod': 'Exp2Syn', 'tau1': 0.8, 'tau2': 5.3, 'e': 0}  # NMDA synaptic mechanism
-netParams.synMechParams['inh'] = {'mod': 'Exp2Syn', 'tau1': 0.6, 'tau2': 8.5, 'e': -75}  # GABA synaptic mechanism
 
 
-# Stimulation parameters
-netParams.stimSourceParams['bkg'] = {'type': 'NetStim', 'rate': 20, 'noise': 0.3}
-netParams.stimTargetParams['bkg->all'] = {'source': 'bkg', 'conds': {'cellType': ['E','I']}, 'weight': 0.01, 'delay': 'max(1, normal(5,2))', 'synMech': 'exc'}
+# ## Cell connectivity rules
+# netParams.connParams['E->all'] = {
+#   'preConds': {'cellType': 'S2'}, 'postConds': {'pop': 'E2'},  # S2->E2
+#   'convergence': 10 ,                  # probability of connection
+#   'weight': '0.005*post_ynorm',         # synaptic weight 
+#   'delay': 'dist_3D/propVelocity',      # transmission delay (ms) 
+#   'synMech': 'exc'}                     # synaptic mechanism 
 
-
-## Cell connectivity rules
-netParams.connParams['E->all'] = {
-  'preConds': {'cellType': 'E'}, 'postConds': {'y': [100,1000]},  #  E -> all (100-1000 um)
-  'probability': 0.1 ,                  # probability of connection
-  'weight': '0.005*post_ynorm',         # synaptic weight 
-  'delay': 'dist_3D/propVelocity',      # transmission delay (ms) 
-  'synMech': 'exc'}                     # synaptic mechanism 
-
-netParams.connParams['I->E'] = {
-  'preConds': {'cellType': 'I'}, 'postConds': {'pop': ['E2','E4','E5']},       #  I -> E
-  'probability': '0.4*exp(-dist_3D/probLengthConst)',   # probability of connection
-  'weight': 0.001,                                      # synaptic weight 
-  'delay': 'dist_3D/propVelocity',                      # transmission delay (ms) 
-  'synMech': 'inh'}                                     # synaptic mechanism 
+#                             # synaptic mechanism 
 
 
 # Simulation options
-simConfig = specs.SimConfig()        # object of class SimConfig to store simulation configuration
-simConfig.duration = 1*1e3           # Duration of the simulation, in ms
-simConfig.dt = 0.025                # Internal integration timestep to use
-simConfig.verbose = False            # Show detailed messages 
-simConfig.recordTraces = {'V_soma':{'sec':'soma','loc':0.5,'var':'v'}}  # Dict with traces to record
-simConfig.recordStep = 1             # Step size in ms to save data (eg. V traces, LFP, etc)
-simConfig.filename = 'tut5'  # Set file output name
-simConfig.savePickle = False         # Save params, network and sim output to pickle file
-simConfig.saveMat = False         # Save params, network and sim output to pickle file
-simConfig.saveJson=1
+cfg = specs.SimConfig()        # object of class cfg to store simulation configuration
+cfg.duration = 5*1e3           # Duration of the simulation, in ms
+cfg.dt = 0.025                # Internal integration timestep to use
+cfg.verbose = False            # Show detailed messages 
+cfg.recordTraces = {'V_soma':{'sec':'soma','loc':0.5,'var':'v'}}  # Dict with traces to record
+cfg.recordStep = 1             # Step size in ms to save data (eg. V traces, LFP, etc)
+cfg.filename = 'ssc-3_sim'  # Set file output name
+cfg.savePickle = False         # Save params, network and sim output to pickle file
+cfg.saveMat = False         # Save params, network and sim output to pickle file
+cfg.saveJson=1
 
-#simConfig.analysis['plotRaster'] = {'orderBy': 'y', 'orderInverse': True}      # Plot a raster
-simConfig.analysis['plotSpikeStats'] = {'include': [['E2','E4'], [15,16]]}      # Plot recorded traces for this list of cells
-#simConfig.analysis['plot2Dnet'] = True            # plot 2D visualization of cell positions and connections
-#simConfig.analysis['plotConn'] = True             # plot connectivity matrix
+cfg.analysis['plotRaster'] = { 'saveFig': True, 'showFig': False, 'labels': 'overlay', 'popRates': True, 'orderInverse': True, 
+							 'figSize': (12,10), 'lw': 0.3, 'markerSize':3, 'marker': '.', 'dpi': 300} 
+
+cfg.analysis['plotSpikeHist'] = {'yaxis':'rate', 'binSize':5, 'graphType':'bar',
+								'saveFig': True, 'showFig': False, 'popColors': popColors, 'figSize': (10,4), 'dpi': 300} 
+cfg.analysis['plotSpikeStats'] = {'saveFig': True}
+cfg.analysis['plotRatePSD'] = {'saveFig': True}
+
 
 # Create network and run simulation
 sim.createSimulateAnalyze(netParams = netParams, simConfig = simConfig)