+++ examples/saveLoadV1/create_save.py
@@ -2,5 +2,6 @@ from netpyne import sim
 import params
 
 # Create network and save
-sim.create(netParams=params.netParams, simConfig=params.simConfig)  
+sim.create(netParams=params.netParams, simConfig=params.simConfig)
+sim.gatherData()
 sim.saveData()